Created textConversion.tsx, textConversion.py, and message.txt
Pushed just to have on my laptop.

Note:
I've researched python and javascript libraries and found that they're not 100% compatible with converting latex so we will most likely have to use pandoc on the backend

To Do:
Work on an API Fetch with pandoc so that we can convert the markdown and latex to something readable.
( Will continue to research on this today )